Revert: add: move the sentry functionality to separate file to avoid cycle dependencies of analytics.js - pino-logger.js files, extend the error method of pino-logger to use 3rd party apis
