+++ src/init.js
@@ -4,7 +4,6 @@ import goodWallet from './lib/wallet/GoodWallet'
 import userStorage from './lib/gundb/UserStorage'
 import isWebApp from './lib/utils/isWebApp'
 import { APP_OPEN, fireEvent, initAnalytics } from './lib/analytics/analytics'
-import { initSentry } from './lib/analytics/sentry'
 import { extractQueryParams } from './lib/share'
 import { setUserStorage, setWallet } from './lib/undux/SimpleStore'
 
@@ -17,7 +16,6 @@ export const init = () => {
 
     // set userStorage to simple storage
     setUserStorage(userStorage)
-    await initSentry(goodWallet, userStorage)
     await initAnalytics(goodWallet, userStorage)
 
     const params = extractQueryParams(window.location.href)
