+++ src/lib/analytics/analytics.js
@@ -1,5 +1,7 @@
 //@flow
 import _debounce from 'lodash/debounce'
+import _forEach from 'lodash/forEach'
+import * as Sentry from '@sentry/browser'
 import logger from '../../lib/logger/pino-logger'
 import Config from '../../config/config'
 
@@ -89,6 +91,21 @@ export const initAnalytics = async (goodWallet: GoodWallet, userStorage: UserSto
     }
   }
 
+  Sentry.init({
+    dsn: Config.sentryDSN,
+  })
+
+  Sentry.configureScope(scope => {
+    if (email || identifier) {
+      scope.setUser({
+        id: identifier,
+        email: email,
+      })
+    }
+
+    scope.setTag('appVersion', Config.version)
+  })
+
   log.debug('Initialized analytics:', {
     Amplitude: Amplitude !== undefined,
     FS: FS !== undefined,
@@ -97,6 +114,21 @@ export const initAnalytics = async (goodWallet: GoodWallet, userStorage: UserSto
 
   patchLogger()
 }
+
+export const reportToSentry = (error, extra = {}, tags = {}) =>
+  Sentry.configureScope(scope => {
+    // set extra
+    _forEach(extra, (value, key) => {
+      scope.setExtra(key, value)
+    })
+
+    // set tags
+    _forEach(tags, (value, key) => {
+      scope.setTags(key, value)
+    })
+
+    Sentry.captureException(error)
+  })
 
 export const fireEvent = (event: string, data: any = {}) => {
   if (Amplitude === undefined) {
